Several smaller improvements for EDS HTML Full Text Rendering (#4347)
* Empower sites to hide the ANID string

EDS API HTML FT will contain a string on top that contains various details such as Accession Number & dates. This string might be something that some sides would rather hide.

The toHTML function converts the <anid> tag into a <p> tag. Adding the "eds_html_anid" class, will allow sites to hide this string via CSS, if they chose to do so

* render tables as tables rather than raw html

Tables would show as raw html, this attempts to split the string at ephtml and if the text starts with table it'll be wrapped in a div with class eds_html_table and the data will be run through html_entity_decode

We only bother if the string contains ephtml

* Ensuring mathematical annotations show

this will help mathematical annotations show, if they are created in MathML and if your browser supports it