Add a small delay to Demo driver's payment registration. (#4654)
This avoids a potential race condition with session writes during CI tests where the "main" process shutdown may be delayed due to code coverage processing causing session writes to happen in reverse order and fines list to not be updated properly.
This also makes the process feel more realistic during manual testing, since it typically takes a moment for an ILS to register a payment.